OPEN - task 2: Invoking the advance_status service results in a 403 Forbidden https://github.com/cedadev/djcharme/issues/issue/2
Decorates with @csrf_exempt the advance_status service.
Refactores partially the unit tests.

+++ djcharme/djcharme/test/http_requests.py
@@ -5,7 +5,7 @@ Created on 25 Jul 2013
 '''
 
 
-from djcharme.views.node_gate import insert, advance_status, process_page
+from djcharme.views.node_gate import insert, process_page
 from djcharme.charme_middleware import CharmeMiddleware
 from djcharme.test import turtle_usecase1
 from django.contrib.auth.models import User
@@ -14,10 +14,8 @@ from django.test.client import RequestFactory
 from rdflib.graph import Graph
 
 import unittest
-import json
 import logging
 from djcharme import settings
-from xml.etree import ElementTree
 
 LOGGING = logging.getLogger(__name__)
 
@@ -41,17 +39,8 @@ class Test(unittest.TestCase):
     def tearDown(self):   
         for res in self.g:
             self.g.remove(res)
-        self.user.delete()
-        
-
-    def extract_annotation_uri(self, document):
-        xml = ElementTree.fromstring(document)
-        RDF = "{http://www.w3.org/1999/02/22-rdf-syntax-ns#}"
-        descriptions = xml.findall('%sDescription' % (RDF))
-        for desc in descriptions:
-            anno = desc.find('./%stype[@%sresource="http://www.w3.org/ns/oa#Annotation"]' % (RDF, RDF))
-            if anno is not None:
-                return desc.get('%sabout' % RDF)
+        if hasattr(self, 'user'):
+            self.user.delete()
 
     def test_insert_anotation(self):
         response = insert(self.factory.post('/insert/annotation',
@@ -101,28 +90,7 @@ class Test(unittest.TestCase):
         self.assert_('hasTarget' not in str(index(request, graph)), 
                      "Cannot generate index page")
     '''
-    def test_advance_status(self):
-        new_state = 'stable' 
-        response = self.test_insert_anotation()
 
-        start = response.content.rfind('<', 0, response.content.index("oa:Annotation")) + 1
-        end = response.content.rfind('>', start, response.content.index("oa:Annotation"))
-
-        data = {'annotation': response.content[start:end], 
-                'toState': new_state}        
-       
-        
-        response = advance_status(self.factory.post('/advance_status',
-                                            content_type='application/json',
-                                            data=json.dumps(data))) 
-        
-        '''
-            Need to verify better the triple has been moved
-            for example using the  (TBD)
-        '''
-        
-        self.assert_(response.status_code == 200, "HTTPResponse has status_code: %s" % response.status_code)
-        #self.assert_(('%s a rdfg:Graph' % (new_state)) in response.content, "Response content does not return the correct state")
         
 
 if __name__ == "__main__":
